feat: add transformSelection tests and utility functions
- Introduced new tests for remote selection handling in collaborative editing, covering various scenarios of text replacement and selection state.
- Added utility functions for managing expected selection states and text transformations, enhancing the testing framework for selection behavior.
- Updated existing utility exports to support the new functionality, ensuring consistency across the editor's selection management.
